Introduced RuptureImporter [skip hazardlib]
